@@ -31,22 +31,23 @@
     "php": ">=8.3",
     "ext-json": "*",
     "beberlei/assert": "^3.3",
-    "behat/behat": "^3.8",
-    "firebase/php-jwt": "^6.4",
-    "guzzlehttp/guzzle": "^7.3"
+    "behat/behat": "^3.29",
+    "firebase/php-jwt": "^6.11",
+    "guzzlehttp/guzzle": "^7.10"
   },
   "require-dev": {
-    "friendsofphp/php-cs-fixer": "^3.70",
-    "imbo/imbo-coding-standard": "^2.0",
+    "friendsofphp/php-cs-fixer": "^3.92",
+    "imbo/imbo-coding-standard": "^3.0",
     "phpstan/extension-installer": "^1.4",
     "phpstan/phpstan": "^2.1",
     "phpstan/phpstan-deprecation-rules": "^2.0",
     "phpstan/phpstan-phpunit": "^2.0",
-    "phpunit/phpunit": "^12.0",
-    "slim/psr7": "^1.3",
-    "slim/slim": "^4.7",
-    "symfony/process": "^7.2",
-    "tuupola/slim-basic-auth": "^3.3"
+    "phpunit/phpunit": "^12.5",
+    "slim/psr7": "^1.8",
+    "slim/slim": "^4.15",
+    "symfony/process": "^7.4",
+    "symfony/var-dumper": "^7.4",
+    "tuupola/slim-basic-auth": "^3.4"
   },
   "autoload": {
     "psr-4": {
@@ -59,23 +60,24 @@
     }
   },
   "scripts": {
+    "behat": "vendor/bin/behat --strict",
     "ci": [
-      "@phpunit",
       "@behat",
-      "@sa",
-      "@cs"
-    ],
-    "test": [
+      "@cs",
       "@phpunit",
-      "@behat"
+      "@sa"
     ],
+    "cs": "vendor/bin/php-cs-fixer check --diff",
+    "cs:fix": "vendor/bin/php-cs-fixer fix --diff",
+    "dev": "php -S localhost:8080 -t ./features/bootstrap",
+    "docs": "cd docs; make html",
     "phpunit": "vendor/bin/phpunit --colors=always",
     "phpunit:coverage": "@phpunit -- --coverage-html build/coverage",
-    "behat": "vendor/bin/behat --strict --colors",
     "sa": "vendor/bin/phpstan",
-    "cs": "vendor/bin/php-cs-fixer fix --dry-run --diff",
-    "dev": "php -S localhost:8080 -t ./features/bootstrap",
-    "docs": "cd docs; make html"
+    "test": [
+      "@behat",
+      "@phpunit"
+    ]
   },
   "config": {
     "sort-packages": true,
