@@ -1,4 +1,5 @@
 <?php declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\Context\Initializer;
 
 use Imbo\BehatApiExtension\ArrayContainsComparator;
@@ -19,18 +20,17 @@ class ArrayContainsComparatorAwareInitializerTest extends TestCase
             ->expects($this->exactly(8))
             ->method('addFunction')
             ->willReturnCallback(
-                fn (string $matcher, object $impl) =>
-                    match ([$matcher, get_class($impl)]) {
-                        ['arrayLength', Matcher\ArrayLength::class]       => $comparator,
-                        ['arrayMinLength', Matcher\ArrayMinLength::class] => $comparator,
-                        ['arrayMaxLength', Matcher\ArrayMaxLength::class] => $comparator,
-                        ['variableType', Matcher\VariableType::class]     => $comparator,
-                        ['regExp', Matcher\RegExp::class]                 => $comparator,
-                        ['gt', Matcher\GreaterThan::class]                => $comparator,
-                        ['lt', Matcher\LessThan::class]                   => $comparator,
-                        ['jwt', Matcher\JWT::class]                       => $comparator,
-                        default                                           => $this->fail("Unexpected matcher: " . $matcher)
-                    },
+                fn (string $matcher, object $impl) => match ([$matcher, $impl::class]) {
+                    ['arrayLength', Matcher\ArrayLength::class] => $comparator,
+                    ['arrayMinLength', Matcher\ArrayMinLength::class] => $comparator,
+                    ['arrayMaxLength', Matcher\ArrayMaxLength::class] => $comparator,
+                    ['variableType', Matcher\VariableType::class] => $comparator,
+                    ['regExp', Matcher\RegExp::class] => $comparator,
+                    ['gt', Matcher\GreaterThan::class] => $comparator,
+                    ['lt', Matcher\LessThan::class] => $comparator,
+                    ['jwt', Matcher\JWT::class] => $comparator,
+                    default => $this->fail('Unexpected matcher: '.$matcher),
+                },
             );
 
         new ArrayContainsComparatorAwareInitializer($comparator);
