@@ -1,12 +1,9 @@
 <?php declare(strict_types=1);
-
 require 'vendor/autoload.php';
 
-$finder = (new Symfony\Component\Finder\Finder())
-    ->files()
-    ->name('*.php')
-    ->in(__DIR__)
-    ->exclude('vendor');
+$finder = new PhpCsFixer\Finder();
+$config = new Imbo\CodingStandard\Config();
 
-return (new Imbo\CodingStandard\Config())
-    ->setFinder($finder);
+return $config
+    ->setParallelConfig(PhpCsFixer\Runner\Parallel\ParallelConfigFactory::detect())
+    ->setFinder($finder->in(__DIR__)->append([__FILE__]));
