@@ -1,42 +1,36 @@
 <?php declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
 use InvalidArgumentException;
 
+use function sprintf;
+
 /**
- * Check if a numeric value is greater than another value
+ * Check if a numeric value is greater than another value.
  */
 class GreaterThan
 {
     /**
-     * Match a numeric value
+     * Match a numeric value.
      *
      * @param mixed $number A variable
-     * @param mixed $min The minimum value of $number
+     * @param mixed $min    The minimum value of $number
+     *
      * @throws InvalidArgumentException
      */
     public function __invoke(mixed $number, mixed $min): bool
     {
         if (!is_numeric($number)) {
-            throw new InvalidArgumentException(sprintf(
-                '"%s" is not numeric.',
-                (string) $number,
-            ));
+            throw new InvalidArgumentException(sprintf('"%s" is not numeric.', (string) $number));
         }
 
         if (!is_numeric($min)) {
-            throw new InvalidArgumentException(sprintf(
-                '"%s" is not numeric.',
-                (string) $min,
-            ));
+            throw new InvalidArgumentException(sprintf('"%s" is not numeric.', (string) $min));
         }
 
         if ($number <= $min) {
-            throw new InvalidArgumentException(sprintf(
-                '"%s" is not greater than "%s".',
-                $number,
-                $min,
-            ));
+            throw new InvalidArgumentException(sprintf('"%s" is not greater than "%s".', $number, $min));
         }
 
         return true;
