@@ -1,4 +1,5 @@
 <?php declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\Context\Initializer;
 
 use Behat\Behat\Context\Context;
@@ -6,7 +7,7 @@ use Behat\Behat\Context\Initializer\ContextInitializer;
 use Imbo\BehatApiExtension\Context\ApiClientAwareContext;
 
 /**
- * API client aware initializer
+ * API client aware initializer.
  *
  * Initializer for feature contexts that implements the ApiClientAwareContext interface.
  */
@@ -14,12 +15,13 @@ class ApiClientAwareInitializer implements ContextInitializer
 {
     /**
      * @var array<mixed> Guzzle client configuration array
+     *
      * @see http://docs.guzzlephp.org/ Check out the Guzzle docs for a complete overview of available configuration parameters
      */
     private array $config = [];
 
     /**
-     * Class constructor
+     * Class constructor.
      *
      * @param array<mixed> $config Guzzle client configuration array
      */
@@ -29,7 +31,7 @@ class ApiClientAwareInitializer implements ContextInitializer
     }
 
     /**
-     * Initialize the context
+     * Initialize the context.
      *
      * Inject the Guzzle client if the context implements the ApiClientAwareContext interface
      */
