@@ -1,10 +1,11 @@
 <?php declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\Exception;
 
 use Exception;
 
 /**
- * Assertion failure exception
+ * Assertion failure exception.
  */
 class AssertionFailedException extends Exception
 {
