@@ -1,4 +1,5 @@
 <?php declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension;
 
 use Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
@@ -13,7 +14,7 @@ class ArrayContainsComparatorTest extends TestCase
 {
     private ArrayContainsComparator $comparator;
 
-    public function setUp(): void
+    protected function setUp(): void
     {
         $this->comparator = new ArrayContainsComparator();
     }
@@ -60,8 +61,7 @@ class ArrayContainsComparatorTest extends TestCase
                                 [
                                     [
                                         [
-                                            'new-list' =>
-                                            [
+                                            'new-list' => [
                                                 8,
                                             ],
                                         ],
@@ -98,8 +98,7 @@ class ArrayContainsComparatorTest extends TestCase
                                     [
                                         7,
                                         [
-                                            'new-list' =>
-                                            [
+                                            'new-list' => [
                                                 8,
                                             ],
                                         ],
@@ -293,22 +292,22 @@ class ArrayContainsComparatorTest extends TestCase
                     'foo' => 'bar',
                 ],
                 'exceptionMessage' => <<<'EXCEPTION'
-The element at key "foo" in the haystack object is not a list.
+                The element at key "foo" in the haystack object is not a list.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "foo[0]": "bar"
-}
+                ================================================================================
+                = Needle =======================================================================
+                ================================================================================
+                {
+                    "foo[0]": "bar"
+                }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "foo": "bar"
-}
-EXCEPTION
+                ================================================================================
+                = Haystack =====================================================================
+                ================================================================================
+                {
+                    "foo": "bar"
+                }
+                EXCEPTION,
             ],
             'an object' => [
                 'needle' => [
@@ -320,24 +319,24 @@ EXCEPTION
                     ],
                 ],
                 'exceptionMessage' => <<<'EXCEPTION'
-The element at key "foo" in the haystack object is not a list.
+                The element at key "foo" in the haystack object is not a list.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "foo[0]": "bar"
-}
+                ================================================================================
+                = Needle =======================================================================
+                ================================================================================
+                {
+                    "foo[0]": "bar"
+                }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "foo": {
-        "foo": "bar"
-    }
-}
-EXCEPTION
+                ================================================================================
+                = Haystack =====================================================================
+                ================================================================================
+                {
+                    "foo": {
+                        "foo": "bar"
+                    }
+                }
+                EXCEPTION,
             ],
         ];
     }
@@ -514,24 +513,24 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-The needle is a list, while the haystack is not.
+            The needle is a list, while the haystack is not.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-[
-    1,
-    2,
-    3
-]
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            [
+                1,
+                2,
+                3
+            ]
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "foo": "bar"
-}
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            {
+                "foo": "bar"
+            }
+            EXCEPTION,
         );
         $this->comparator->compare([1, 2, 3], ['foo' => 'bar']);
     }
@@ -550,22 +549,22 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-Haystack does not contain any list elements, needle can't be found.
+            Haystack does not contain any list elements, needle can't be found.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-[
-    1
-]
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            [
+                1
+            ]
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-[
-    1
-]
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            [
+                1
+            ]
+            EXCEPTION,
         );
         $this->comparator->compare([[1]], [1]);
     }
@@ -575,26 +574,26 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-Haystack does not contain any object elements, needle can't be found.
+            Haystack does not contain any object elements, needle can't be found.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "foo": "bar"
-}
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            {
+                "foo": "bar"
+            }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-[
-    [
-        1,
-        2,
-        3
-    ]
-]
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            [
+                [
+                    1,
+                    2,
+                    3
+                ]
+            ]
+            EXCEPTION,
         );
         $this->comparator->compare([
             ['foo' => 'bar'],
@@ -608,34 +607,34 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-The list in needle was not found in the list elements in the haystack.
+            The list in needle was not found in the list elements in the haystack.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-[
-    1,
-    3
-]
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            [
+                1,
+                3
+            ]
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-[
-    [
-        1,
-        2
-    ],
-    [
-        3,
-        4
-    ],
-    [
-        5,
-        6
-    ]
-]
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            [
+                [
+                    1,
+                    2
+                ],
+                [
+                    3,
+                    4
+                ],
+                [
+                    5,
+                    6
+                ]
+            ]
+            EXCEPTION,
         );
         $this->comparator->compare([
             [1, 3],
@@ -651,37 +650,37 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-The object in needle was not found in the object elements in the haystack.
+            The object in needle was not found in the object elements in the haystack.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "id": 2,
-    "gameId": 3
-}
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            {
+                "id": 2,
+                "gameId": 3
+            }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-[
-    {
-        "id": 1,
-        "gameId": 1,
-        "value": 1
-    },
-    {
-        "id": 2,
-        "gameId": 2,
-        "value": 2
-    },
-    {
-        "id": 3,
-        "gameId": 3,
-        "value": 3
-    }
-]
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            [
+                {
+                    "id": 1,
+                    "gameId": 1,
+                    "value": 1
+                },
+                {
+                    "id": 2,
+                    "gameId": 2,
+                    "value": 2
+                },
+                {
+                    "id": 3,
+                    "gameId": 3,
+                    "value": 3
+                }
+            ]
+            EXCEPTION,
         );
         $this->comparator->compare([
             [
@@ -723,22 +722,22 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-Haystack object is missing the "bar" key.
+            Haystack object is missing the "bar" key.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "bar": "baz"
-}
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            {
+                "bar": "baz"
+            }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "baz": "bar"
-}
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            {
+                "baz": "bar"
+            }
+            EXCEPTION,
         );
         $this->comparator->compare([
             'foo' => [
@@ -755,22 +754,22 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-Value mismatch for key "foo" in haystack object.
+            Value mismatch for key "foo" in haystack object.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "foo": "bar"
-}
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            {
+                "foo": "bar"
+            }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "foo": "baz"
-}
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            {
+                "foo": "baz"
+            }
+            EXCEPTION,
         );
         $this->comparator->compare([
             'foo' => 'bar',
@@ -812,24 +811,24 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-Value mismatch for key "foo" in haystack object.
+            Value mismatch for key "foo" in haystack object.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "bar": "foo",
-    "foo": "foobar"
-}
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            {
+                "bar": "foo",
+                "foo": "foobar"
+            }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "bar": "foo",
-    "foo": "foo"
-}
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            {
+                "bar": "foo",
+                "foo": "foo"
+            }
+            EXCEPTION,
         );
         $this->comparator->compare([
             'foo' => [
@@ -862,22 +861,22 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-Haystack object is missing the "foo" key.
+            Haystack object is missing the "foo" key.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "foo[0]": "bar"
-}
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            {
+                "foo[0]": "bar"
+            }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "bar": "foo"
-}
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            {
+                "bar": "foo"
+            }
+            EXCEPTION,
         );
         $this->comparator->compare([
             'foo[0]' => 'bar',
@@ -899,25 +898,25 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-The index "2" does not exist in the haystack list.
+            The index "2" does not exist in the haystack list.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "foo[2]": "bar"
-}
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            {
+                "foo[2]": "bar"
+            }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "foo": [
-        "foo",
-        "bar"
-    ]
-}
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            {
+                "foo": [
+                    "foo",
+                    "bar"
+                ]
+            }
+            EXCEPTION,
         );
         $this->comparator->compare([
             'foo[2]' => 'bar',
@@ -934,18 +933,18 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-Value mismatch for index "0" in haystack list.
+            Value mismatch for index "0" in haystack list.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-"foo"
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            "foo"
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-"foobar"
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            "foobar"
+            EXCEPTION,
         );
         $this->comparator->compare([
             '[0]' => 'foo',
@@ -970,22 +969,22 @@ EXCEPTION
         $this->expectException(ArrayContainsComparatorException::class);
         $this->expectExceptionMessage(
             <<<'EXCEPTION'
-Value mismatch for key "key" in haystack object.
+            Value mismatch for key "key" in haystack object.
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "key": "@foo(123)"
-}
+            ================================================================================
+            = Needle =======================================================================
+            ================================================================================
+            {
+                "key": "@foo(123)"
+            }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "key": "some value"
-}
-EXCEPTION
+            ================================================================================
+            = Haystack =====================================================================
+            ================================================================================
+            {
+                "key": "some value"
+            }
+            EXCEPTION,
         );
         $this->comparator->compare(
             ['key' => '@foo(123)'],
