@@ -1,4 +1,5 @@
 <?php declare(strict_types=1);
+
 namespace Imbo\BehatApiExtension\Exception;
 
 use PHPUnit\Framework\Attributes\CoversClass;
@@ -19,40 +20,40 @@ class ArrayContainsComparatorExceptionTest extends TestCase
                 'needle' => [],
                 'haystack' => [],
                 'formattedMessage' => <<<MESSAGE
-{$someMessage}
+                {$someMessage}
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-[]
+                ================================================================================
+                = Needle =======================================================================
+                ================================================================================
+                []
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-[]
-MESSAGE
+                ================================================================================
+                = Haystack =====================================================================
+                ================================================================================
+                []
+                MESSAGE,
             ],
             'with needle and haystack' => [
                 'message' => $someMessage = 'some message',
                 'needle' => ['needle' => 'value'],
                 'haystack' => ['haystack' => 'value'],
                 'formattedMessage' => <<<MESSAGE
-{$someMessage}
+                {$someMessage}
 
-================================================================================
-= Needle =======================================================================
-================================================================================
-{
-    "needle": "value"
-}
+                ================================================================================
+                = Needle =======================================================================
+                ================================================================================
+                {
+                    "needle": "value"
+                }
 
-================================================================================
-= Haystack =====================================================================
-================================================================================
-{
-    "haystack": "value"
-}
-MESSAGE
+                ================================================================================
+                = Haystack =====================================================================
+                ================================================================================
+                {
+                    "haystack": "value"
+                }
+                MESSAGE,
             ],
         ];
     }
